+++ docs/misc/changelog.rst
@@ -11,9 +11,11 @@ Pre Release 2.0.1.a0 (WIP)
 **logging and bug fixes**
 
 - added patch fix for equal function using `gym.spaces.MultiDiscrete` and `gym.spaces.MultiBinary`
+- fixes for DQN action_probability
+- re-added double DQN + refactored DQN policies **breaking changes**
+- replaced `async` with `async_eigen_decomp` in ACKTR/KFAC for python 3.7 compat
 - removed action clipping for prediction of continuous actions (see issue #36)
 - fixed NaN issue due to clipping the continuous action in the wrong place (issue #36)
-
 
 Release 2.0.0 (2018-09-18)
 --------------------------